+++ LotusCatering/Web/LotusCatering.Web.ViewModels/Items/ItemViewModel.cs
@@ -15,10 +15,8 @@
 
         public string ImageUrl { get; set; }
 
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
         public string Description { get; set; }
-
-        // public string UserCartId { get; set; }
     }
 }
